Move service container to services folder
